Moved to_Fourier from to_vmec.py to util.py, since it is now used for plotting also. Removed the theta parameter so theta is treated the same as phi
