+++ qsc/util.py
@@ -62,6 +62,60 @@ def fourier_minimum(y):
     solution = scipy.optimize.minimize_scalar(func, bracket=bracket)
     return solution.fun
 
+def to_Fourier(R_2D, Z_2D, nfp, mpol, ntor, lasym):
+    """
+    This function takes two 2D arrays (R_2D and Z_2D), which contain
+    the values of the radius R and vertical coordinate Z in cylindrical
+    coordinates of a given surface and Fourier transform it, outputing
+    the resulting cos(theta) and sin(theta) Fourier coefficients
+
+    The first dimension of R_2D and Z_2D should correspond to the
+    theta grid, while the second dimension should correspond to the
+    phi grid.
+
+    Args:
+        R_2D: 2D array of the radial coordinate R(theta, phi) of a given surface
+        Z_2D: 2D array of the vertical coordinate Z(theta, phi) of a given surface
+        nfp: number of field periods of the surface
+        mpol: resolution in poloidal Fourier space
+        ntor: resolution in toroidal Fourier space
+        lasym: False if stellarator-symmetric, True if not
+    """
+    shape = np.array(R_2D).shape
+    ntheta = shape[0]
+    nphi_conversion = shape[1]
+    theta = np.linspace(0, 2 * np.pi, ntheta, endpoint=False)
+    phi_conversion = np.linspace(0, 2 * np.pi / nfp, nphi_conversion, endpoint=False)
+    RBC = np.zeros((int(2 * ntor + 1), int(mpol + 1)))
+    RBS = np.zeros((int(2 * ntor + 1), int(mpol + 1)))
+    ZBC = np.zeros((int(2 * ntor + 1), int(mpol + 1)))
+    ZBS = np.zeros((int(2 * ntor + 1), int(mpol + 1)))
+    factor = 2 / (ntheta * nphi_conversion)
+    phi2d, theta2d = np.meshgrid(phi_conversion, theta)
+    for m in range(mpol+1):
+        nmin = -ntor
+        if m==0: nmin = 1
+        for n in range(nmin, ntor+1):
+            angle = m * theta2d - n * nfp * phi2d
+            sinangle = np.sin(angle)
+            cosangle = np.cos(angle)
+            factor2 = factor
+            # The next 2 lines ensure inverse Fourier transform(Fourier transform) = identity
+            if np.mod(ntheta,2) == 0 and m  == (ntheta/2): factor2 = factor2 / 2
+            if np.mod(nphi_conversion,2) == 0 and abs(n) == (nphi_conversion/2): factor2 = factor2 / 2
+            RBC[n + ntor, m] = np.sum(R_2D * cosangle * factor2)
+            RBS[n + ntor, m] = np.sum(R_2D * sinangle * factor2)
+            ZBC[n + ntor, m] = np.sum(Z_2D * cosangle * factor2)
+            ZBS[n + ntor, m] = np.sum(Z_2D * sinangle * factor2)
+    RBC[ntor,0] = np.sum(R_2D) / (ntheta * nphi_conversion)
+    ZBC[ntor,0] = np.sum(Z_2D) / (ntheta * nphi_conversion)
+
+    if not lasym:
+        RBS = 0
+        ZBC = 0
+
+    return RBC, RBS, ZBC, ZBS
+
 def B_mag(self, r, theta, phi, Boozer_toroidal = False):
     '''
     Function to calculate the modulus of the magnetic field B for a given
